Finalize macOS installation

+++ install/linux/install_appium_inspector.sh
@@ -8,27 +8,40 @@ echo "Setting up Appium Inspector..."
 APPIUM_INSPECTOR_DIR="$HOME/.appium-inspector"
 mkdir -p $APPIUM_INSPECTOR_DIR
 
-# Download Appium Inspector for Linux
-APPIUM_INSPECTOR_URL="https://github.com/appium/appium-inspector/releases/download/v2024.12.1/Appium-Inspector-2024.12.1-linux-x86_64.AppImage"
-APPIUM_INSPECTOR_FILE="$APPIUM_INSPECTOR_DIR/appium-inspector.AppImage"
-
-echo "Downloading Appium Inspector for Linux..."
-curl -L $APPIUM_INSPECTOR_URL -o $APPIUM_INSPECTOR_FILE
-chmod +x $APPIUM_INSPECTOR_FILE
-
 # Create launcher script
 USER_BIN_DIR="$HOME/bin"
-mkdir -p $USER_BIN_DIR
+mkdir -p "$USER_BIN_DIR"
 APPIUM_INSPECTOR_SCRIPT="$USER_BIN_DIR/appium_inspector"
+
+if [ -f "$APPIUM_INSPECTOR_SCRIPT" ]; then
+        echo "Appium Inspector is already installed."
+        exit 0
+fi
+
+if [ -f "$CURRENT_DIR/Appium-inspector.dmg" ]; then
+   cp "$CURRENT_DIR/Appium-inspector.dmg" APPIUM_INSPECTOR_FILE
+else
+  # Download Appium Inspector for Linux
+  APPIUM_INSPECTOR_URL="https://github.com/appium/appium-inspector/releases/download/v2024.12.1/Appium-Inspector-2024.12.1-linux-x86_64.AppImage"
+  APPIUM_INSPECTOR_FILE="$APPIUM_INSPECTOR_DIR/appium-inspector.AppImage"
+
+  echo "Downloading Appium Inspector for Linux..."
+  curl -L $APPIUM_INSPECTOR_URL -o "$APPIUM_INSPECTOR_FILE"
+  chmod +x "$APPIUM_INSPECTOR_FILE"
+fi
+
 
 cat > $APPIUM_INSPECTOR_SCRIPT << EOF
 #!/bin/bash
 $APPIUM_INSPECTOR_FILE "\$@"
 EOF
-chmod +x $APPIUM_INSPECTOR_SCRIPT
+chmod +x "$APPIUM_INSPECTOR_SCRIPT"
+
+source "$CURRENT_DIR"/../common/get_shell.sh
+SHELL_PROFILE=$(shell_profile)
 
 # Add to PATH if not already there
-if ! echo $PATH | grep -q "$USER_BIN_DIR"; then
-    echo "export PATH=\$PATH:$USER_BIN_DIR" >> $SHELL_PROFILE #TODO this is out of scope? Use get_shell from macos
+if ! echo "$PATH" | grep -q "$USER_BIN_DIR"; then
+    echo "export PATH=\$PATH:$USER_BIN_DIR" >> "$SHELL_PROFILE"
     export PATH=$PATH:$USER_BIN_DIR
 fi
